requirements.txt
Django==5.0.7
django-phone-field==1.8.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
pillow==10.4.0
